requirements.txt
torch==2.2.2
torchvision==0.17.2
timm==0.9.12
pytorch-grad-cam==1.4.1
opencv-python==4.9.0.80
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
matplotlib==3.8.4
seaborn==0.13.2
tqdm==4.66.4
pillow==10.3.0
